requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Dependency Injection (commented out due to build issues)
# dependency-injector==4.41.0

# JSON logging (optional)
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
